Refactor navigation utilities into submodules
