+++ tests/test_log_columns.py
@@ -8,7 +8,7 @@ import tests.conftest  # ensure stubs loaded
 def test_setup_environment_header_includes_perf(monkeypatch, tmp_path):
     airsim_stub = types.SimpleNamespace(ImageRequest=object, ImageType=object)
     monkeypatch.setitem(sys.modules, "airsim", airsim_stub)
-    nl = importlib.import_module("uav.nav_loop")
+    nl = importlib.import_module("uav.nav_runtime")
     importlib.reload(nl)
 
     monkeypatch.setattr(nl, "start_video_writer_thread", lambda *a, **k: types.SimpleNamespace(join=lambda: None))
